feat: add all middlewares about users

+++ src/modules/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UsersRepository } from './repositories/users.repository';
@@ -19,6 +19,11 @@ export class UsersService {
 
   async findOne(id: string) {
     const user = await this.userRepository.findOne(id);
+
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
     return user;
   }
 
@@ -27,11 +32,22 @@ export class UsersService {
     return user;
   }
   async update(id: string, updateUserDto: UpdateUserDto) {
+    const userFound = await this.userRepository.findOne(id);
+
+    if (!userFound) {
+      throw new NotFoundException('User not found');
+    }
+
     const newUser = await this.userRepository.update(id, updateUserDto);
     return newUser;
   }
 
   async delete(id: string) {
+    const userFound = await this.userRepository.findOne(id);
+
+    if (!userFound) {
+      throw new NotFoundException('User not found');
+    }
     await this.userRepository.delete(id);
   }
 }
